scraper_service/requirements.txt
requests
beautifulsoup4
mysql-connector-python
google-genai
python-dotenv
lxml